07-complex-docs-rag/requirements.txt
fastapi
uvicorn
pydantic
httpx
pdfminer.six
chromadb
sentence-transformers
numpy
